refactor: coverage の除外対象がファイル名で決まる
実バインディングの上でしか動かないモジュールとコマンドとして実行される
ファイルは、これまで vitest.config.mts の coverageExclude に 1 行ずつパスを
足さないと gate から外れなかった。新しく該当するモジュールが増えるたびに
config の編集が要り、書き忘れると gate が落ちる。

`*.live.ts` と `*.entry.ts` の 2 つの接尾辞をパタンとして持たせ、5 ファイルを
その名前に改名した。生成物も `src/routeTree.gen.ts` から `src/**/*.gen.ts` に
広げた。coverageInclude が並べていた 3 つのルートは COVERED_ROOTS に括り、
exclude のパタンからも同じ定数を参照する。

- src/lib/auth/auth-client.ts -> auth-client.live.ts
- src/lib/auth/actions.ts -> actions.live.ts
- src/lib/storage/r2.ts -> r2.live.ts
- src/gateways/user/drizzle-store.ts -> drizzle-store.live.ts
- scripts/check-cursor-rule-mirrors.ts -> check-cursor-rule-mirrors.entry.ts
